fixed 2 off-by-one bugs with string concatentation in the QoreString class, fixed get*BigInt() time functions to return int64 instead of int, fixed a memory leak in exception handling in object_MakeString(), cosmetic fix in exception string in Datasource constructor
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/branches/0.6.2@1533 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ include/qore/QoreNode.h
@@ -162,13 +162,13 @@ class QoreNode : public ReferenceObject
 
 // for getting relative time values or integer values
 DLLEXPORT int getSecZeroInt(class QoreNode *a);
-DLLEXPORT int getSecZeroBigInt(class QoreNode *a);
+DLLEXPORT int64 getSecZeroBigInt(class QoreNode *a);
 DLLEXPORT int getSecMinusOneInt(class QoreNode *a);
-DLLEXPORT int getSecMinusOneBigInt(class QoreNode *a);
+DLLEXPORT int64 getSecMinusOneBigInt(class QoreNode *a);
 DLLEXPORT int getMsZeroInt(class QoreNode *a);
-DLLEXPORT int getMsZeroBigInt(class QoreNode *a);
+DLLEXPORT int64 getMsZeroBigInt(class QoreNode *a);
 DLLEXPORT int getMsMinusOneInt(class QoreNode *a);
-DLLEXPORT int getMsMinusOneBigInt(class QoreNode *a);
+DLLEXPORT int64 getMsMinusOneBigInt(class QoreNode *a);
 DLLEXPORT int getMicroSecZeroInt(class QoreNode *a);
 DLLEXPORT bool is_nothing(class QoreNode *n);
 DLLEXPORT bool is_value(class QoreNode *node);
